refactor(planning): split PlanningService into focused services
Break the monolithic PlanningService into three cohesive services and
extract proposal-building helpers into a pure-function class.

- PlanningService (207 lines) — proposal generation + planning overview
- AssignmentService (203 lines) — assignment/group CRUD
- ConfirmationOrchestrationService (322 lines) — confirmation flow,
  emails, non-responders, admin cash confirmation
- PlanningProposalBuilder — BuildConflicts, BuildUnits,
  BuildPreferencesLookup, IntersectPreferences as pure functions

Deduplicated 3x recipient-resolution and 2x email-send blocks in
ConfirmationOrchestrationService via ResolveRecipient and
SendConfirmationEmailAsync helpers.

Rewired 9 planning endpoints to the new services; DI updated.

Test coverage split to match: AssignmentServiceTests (6 tests) and
ConfirmationOrchestrationServiceTests (4 tests) cover the extracted
services; PlanningServiceTests focuses on planning flow only.

Build clean, 95/95 tests pass.
