+++ backend/CoachOS.API/Endpoints/Planning/DeleteAssignmentEndpoint.cs
@@ -8,7 +8,7 @@ public class DeleteAssignmentEndpoint : IEndpoint
     public void MapEndpoint(IEndpointRouteBuilder app)
     {
         app.MapDelete("/lessonseries/{id:guid}/planning/assignments/{assignmentId:guid}",
-            async (Guid id, Guid assignmentId, IPlanningService service, HttpContext ctx, CancellationToken ct) =>
+            async (Guid id, Guid assignmentId, IAssignmentService service, HttpContext ctx, CancellationToken ct) =>
             {
                 var result = await service.DeleteAssignmentAsync(
                     id, assignmentId, ctx.GetOrganizationId(), ct);
