+++ backend/CoachOS.API/Endpoints/Planning/GetNonRespondersEndpoint.cs
@@ -8,7 +8,7 @@ public class GetNonRespondersEndpoint : IEndpoint
     public void MapEndpoint(IEndpointRouteBuilder app)
     {
         app.MapGet("/lessonseries/{id:guid}/planning/non-responders",
-            async (Guid id, IPlanningService service, HttpContext ctx, CancellationToken ct) =>
+            async (Guid id, IConfirmationOrchestrationService service, HttpContext ctx, CancellationToken ct) =>
             {
                 var result = await service.GetNonRespondersAsync(id, ctx.GetOrganizationId(), ct);
                 return result.IsSuccess ? Results.Ok(result.Value) : result.ToErrorResult();
