+++ backend/CoachOS.Application/Planning/PlanningService.cs
@@ -1,13 +1,8 @@
-using System.Security.Cryptography;
-using System.Text;
-using CoachOS.Application.Configuration;
 using CoachOS.Application.Planning.DTOs;
 using CoachOS.Domain.Entities;
 using CoachOS.Domain.Enums;
 using CoachOS.Domain.Interfaces;
 using CoachOS.Domain.Models;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 
 namespace CoachOS.Application.Planning;
 
@@ -17,12 +12,7 @@ public class PlanningService(
     IEnrollmentGroupRepository enrollmentGroupRepo,
     ITimeSlotPreferenceRepository timeSlotPreferenceRepo,
     IScheduleAssignmentRepository scheduleAssignmentRepo,
-    IAssignmentConfirmationTokenRepository tokenRepo,
-    IPaymentRepository paymentRepo,
-    IUserLookupService userLookup,
-    IEmailService emailService,
-    IOptions<AppOptions> appOptions,
-    ILogger<PlanningService> logger) : IPlanningService
+    IUserLookupService userLookup) : IPlanningService
 {
     public async Task<Result<PlanningOverviewDto>> GenerateProposalAsync(
         Guid seriesId, Guid organizationId, bool force = false, CancellationToken ct = default)
@@ -37,7 +27,6 @@ public class PlanningService(
         if (series.PlanningStatus == PlanningStatus.Planning && !force)
             return await GetPlanningOverviewAsync(seriesId, organizationId, ct);
 
-        // Load all data
         var enrollments = await enrollmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
         var activeEnrollments = enrollments
             .Where(e => e.Status is EnrollmentStatus.Confirmed or EnrollmentStatus.Pending)
@@ -50,12 +39,7 @@ public class PlanningService(
             return Result<PlanningOverviewDto>.Fail(
                 new Error(ErrorCodes.Validation, "Er zijn nog geen tijdslots aangemaakt."));
 
-        // Build preference lookup: enrollmentId -> { slotId -> preference }
-        var prefsByEnrollment = preferences
-            .GroupBy(p => p.EnrollmentId)
-            .ToDictionary(
-                g => g.Key,
-                g => g.ToDictionary(p => p.WeeklyTemplateEntryId, p => p.Preference));
+        var prefsByEnrollment = PlanningProposalBuilder.BuildPreferencesLookup(preferences);
 
         // On force-regenerate, preserve manually locked assignments: exclude their
         // units from the algorithm and reduce the affected slots' remaining capacity.
@@ -79,70 +63,8 @@ public class PlanningService(
                 g => g.Key,
                 g => g.Sum(a => a.EnrollmentGroup?.Members.Count ?? 1));
 
-        // Build enrollment units
-        var units = new List<EnrollmentUnit>();
-        var groupedEnrollmentIds = new HashSet<Guid>();
-
-        foreach (var group in groups)
-        {
-            var memberEnrollments = activeEnrollments
-                .Where(e => e.EnrollmentGroupId == group.Id)
-                .ToList();
-
-            if (memberEnrollments.Count == 0) continue;
-
-            // Skip locked groups: they stay where the admin put them.
-            if (lockedGroupIds.Contains(group.Id))
-            {
-                foreach (var e in memberEnrollments)
-                    groupedEnrollmentIds.Add(e.Id);
-                continue;
-            }
-
-            // Use leader's preferences for the group (members don't submit individual preferences).
-            // Fall back to intersection if multiple members have preferences.
-            var membersWithPrefs = memberEnrollments
-                .Where(e => prefsByEnrollment.ContainsKey(e.Id))
-                .ToList();
-
-            Dictionary<Guid, SlotPreference> groupPrefs;
-            if (membersWithPrefs.Count == 0)
-                groupPrefs = new();
-            else if (membersWithPrefs.Count == 1)
-                groupPrefs = prefsByEnrollment[membersWithPrefs[0].Id];
-            else
-                groupPrefs = IntersectPreferences(
-                    membersWithPrefs.Select(e => prefsByEnrollment[e.Id]).ToList());
-
-            var leaderEnrollment = memberEnrollments.FirstOrDefault(e => e.Id == group.LeaderEnrollmentId);
-            units.Add(new EnrollmentUnit(
-                Id: group.Id,
-                IsGroup: true,
-                GroupId: group.Id,
-                EnrollmentIds: memberEnrollments.Select(e => e.Id).ToList(),
-                StudentNames: memberEnrollments.Select(e => e.StudentName).ToList(),
-                Size: memberEnrollments.Count,
-                IsOpenToGrouping: leaderEnrollment?.IsOpenToGrouping ?? false,
-                Preferences: groupPrefs));
-
-            foreach (var e in memberEnrollments)
-                groupedEnrollmentIds.Add(e.Id);
-        }
-
-        foreach (var enrollment in activeEnrollments.Where(e =>
-            !groupedEnrollmentIds.Contains(e.Id) && !lockedEnrollmentIds.Contains(e.Id)))
-        {
-            var prefs = prefsByEnrollment.GetValueOrDefault(enrollment.Id, new());
-            units.Add(new EnrollmentUnit(
-                Id: enrollment.Id,
-                IsGroup: false,
-                GroupId: null,
-                EnrollmentIds: [enrollment.Id],
-                StudentNames: [enrollment.StudentName],
-                Size: 1,
-                IsOpenToGrouping: enrollment.IsOpenToGrouping,
-                Preferences: prefs));
-        }
+        var (units, _) = PlanningProposalBuilder.BuildUnits(
+            activeEnrollments, groups, prefsByEnrollment, lockedGroupIds, lockedEnrollmentIds);
 
         var slotInfos = slots
             .Select(s => new SlotInfo(
@@ -151,13 +73,11 @@ public class PlanningService(
             .ToList();
         var input = new SchedulingInput(units, slotInfos);
 
-        // Run algorithm
         var result = SchedulingAlgorithm.Generate(input);
 
         // Clear existing proposed assignments (uses ExecuteDelete to avoid tracking conflicts)
         await scheduleAssignmentRepo.RemoveProposedBySeriesAsync(seriesId, organizationId, ct);
 
-        // Persist new assignments
         var newAssignments = result.Assignments.Select(a => new ScheduleAssignment
         {
             OrganizationId = organizationId,
@@ -171,7 +91,6 @@ public class PlanningService(
 
         await scheduleAssignmentRepo.AddRangeAsync(newAssignments, ct);
 
-        // Update planning status
         series.PlanningStatus = PlanningStatus.Planning;
         await lessonSeriesRepo.SaveChangesAsync(ct);
 
@@ -195,11 +114,7 @@ public class PlanningService(
             .Where(e => e.Status is EnrollmentStatus.Confirmed or EnrollmentStatus.Pending)
             .ToList();
 
-        var prefsRawByEnrollment = preferences
-            .GroupBy(p => p.EnrollmentId)
-            .ToDictionary(
-                g => g.Key,
-                g => g.ToDictionary(p => p.WeeklyTemplateEntryId, p => p.Preference));
+        var prefsRawByEnrollment = PlanningProposalBuilder.BuildPreferencesLookup(preferences);
         var prefsByEnrollment = prefsRawByEnrollment
             .ToDictionary(
                 kv => kv.Key,
@@ -275,7 +190,7 @@ public class PlanningService(
             ? assignments.Max(a => a.UpdatedAt == default ? a.CreatedAt : a.UpdatedAt)
             : null;
 
-        var conflicts = BuildConflicts(
+        var conflicts = PlanningProposalBuilder.BuildConflicts(
             activeEnrollments, assignments, series.WeeklyTemplate.ToList(), prefsRawByEnrollment);
 
         return Result<PlanningOverviewDto>.Ok(new PlanningOverviewDto
@@ -288,610 +203,5 @@ public class PlanningService(
             Assignments = assignmentDtos,
             Conflicts = conflicts,
         });
-    }
-
-    private static List<PlanningConflictDto> BuildConflicts(
-        List<Enrollment> activeEnrollments,
-        List<ScheduleAssignment> assignments,
-        List<WeeklyTemplateEntry> slots,
-        Dictionary<Guid, Dictionary<Guid, SlotPreference>> prefsByEnrollment)
-    {
-        var conflicts = new List<PlanningConflictDto>();
-
-        // Assigned enrollment IDs (solo + group members)
-        var assignedEnrollmentIds = assignments
-            .SelectMany(a => a.EnrollmentGroup is not null
-                ? a.EnrollmentGroup.Members.Select(m => m.Id)
-                : a.EnrollmentId.HasValue ? [a.EnrollmentId.Value] : Array.Empty<Guid>())
-            .ToHashSet();
-
-        // 1. No viable slot — unassigned + every slot is Unavailable (or missing from prefs)
-        foreach (var enrollment in activeEnrollments)
-        {
-            if (assignedEnrollmentIds.Contains(enrollment.Id)) continue;
-
-            var prefs = prefsByEnrollment.GetValueOrDefault(enrollment.Id, new());
-            bool hasViable = slots.Any(s =>
-                prefs.TryGetValue(s.Id, out var p) &&
-                (p == SlotPreference.Preferred || p == SlotPreference.Available));
-
-            if (!hasViable)
-                conflicts.Add(new PlanningConflictDto
-                {
-                    EnrollmentId = enrollment.Id,
-                    Type = "no_viable_slot",
-                    Message = $"{enrollment.StudentName} heeft geen beschikbaar tijdslot.",
-                });
-        }
-
-        // 2. Oversubscribed slot — assigned count exceeds MaxStudents
-        foreach (var slot in slots)
-        {
-            var count = assignments
-                .Where(a => a.WeeklyTemplateEntryId == slot.Id)
-                .Sum(a => a.EnrollmentGroup?.Members.Count ?? 1);
-
-            if (count > slot.MaxStudents)
-                conflicts.Add(new PlanningConflictDto
-                {
-                    TimeSlotId = slot.Id,
-                    Type = "oversubscribed",
-                    Message = $"Tijdslot overboekt ({count}/{slot.MaxStudents}).",
-                });
-        }
-
-        return conflicts;
-    }
-
-    public async Task<Result<bool>> DeleteAssignmentAsync(
-        Guid seriesId, Guid assignmentId, Guid organizationId, CancellationToken ct = default)
-    {
-        var assignment = await scheduleAssignmentRepo.GetByIdAsync(assignmentId, organizationId, ct);
-        if (assignment is null || assignment.LessonSerieId != seriesId)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Toewijzing niet gevonden."));
-
-        if (assignment.Status == ScheduleAssignmentStatus.Confirmed)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Bevestigde toewijzingen kunnen niet verwijderd worden."));
-
-        scheduleAssignmentRepo.RemoveRange([assignment]);
-        await scheduleAssignmentRepo.SaveChangesAsync(ct);
-
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<bool>> UpdateAssignmentAsync(
-        Guid seriesId, Guid assignmentId, UpdateAssignmentRequest request,
-        Guid organizationId, CancellationToken ct = default)
-    {
-        var assignment = await scheduleAssignmentRepo.GetByIdAsync(assignmentId, organizationId, ct);
-        if (assignment is null || assignment.LessonSerieId != seriesId)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Toewijzing niet gevonden."));
-
-        if (assignment.Status == ScheduleAssignmentStatus.Confirmed)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Bevestigde toewijzingen kunnen niet verplaatst worden."));
-
-        // Capacity guard on target slot (excluding the assignment being moved)
-        var capacityError = await EnsureSlotCapacityAsync(
-            seriesId, organizationId, request.WeeklyTemplateEntryId,
-            addSize: assignment.EnrollmentGroup?.Members.Count ?? 1,
-            excludeAssignmentId: assignment.Id, ct);
-        if (capacityError is not null)
-            return Result<bool>.Fail(capacityError);
-
-        assignment.WeeklyTemplateEntryId = request.WeeklyTemplateEntryId;
-        assignment.IsLocked = true;
-        await scheduleAssignmentRepo.SaveChangesAsync(ct);
-
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<bool>> CreateAssignmentAsync(
-        Guid seriesId, CreateAssignmentRequest request,
-        Guid organizationId, CancellationToken ct = default)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        if (series is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        var slot = series.WeeklyTemplate.FirstOrDefault(s => s.Id == request.WeeklyTemplateEntryId);
-        if (slot is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Tijdslot niet gevonden."));
-
-        var existingAssignments = await scheduleAssignmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-
-        int addSize;
-        Guid? enrollmentId = null;
-        Guid? groupId = null;
-
-        if (request.GroupId.HasValue)
-        {
-            var group = await enrollmentGroupRepo.GetByIdAsync(request.GroupId.Value, organizationId, ct);
-            if (group is null || group.LessonSerieId != seriesId)
-                return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Groep niet gevonden."));
-
-            if (existingAssignments.Any(a => a.EnrollmentGroupId == request.GroupId))
-                return Result<bool>.Fail(new Error(ErrorCodes.Validation, "Groep is al toegewezen."));
-
-            groupId = group.Id;
-            addSize = group.Members.Count;
-        }
-        else
-        {
-            var enrollment = await enrollmentRepo.GetByIdAsync(request.EnrollmentId!.Value, organizationId, ct);
-            if (enrollment is null || enrollment.LessonSerieId != seriesId)
-                return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Inschrijving niet gevonden."));
-
-            if (existingAssignments.Any(a => a.EnrollmentId == request.EnrollmentId))
-                return Result<bool>.Fail(new Error(ErrorCodes.Validation, "Inschrijving is al toegewezen."));
-
-            enrollmentId = enrollment.Id;
-            addSize = 1;
-        }
-
-        var capacityError = await EnsureSlotCapacityAsync(
-            seriesId, organizationId, request.WeeklyTemplateEntryId, addSize, excludeAssignmentId: null, ct);
-        if (capacityError is not null)
-            return Result<bool>.Fail(capacityError);
-
-        ScheduleAssignment assignment = new()
-        {
-            OrganizationId = organizationId,
-            LessonSerieId = seriesId,
-            WeeklyTemplateEntryId = request.WeeklyTemplateEntryId,
-            EnrollmentId = enrollmentId,
-            EnrollmentGroupId = groupId,
-            Status = ScheduleAssignmentStatus.Proposed,
-            IsAutoMerged = false,
-            IsLocked = true,
-        };
-
-        await scheduleAssignmentRepo.AddRangeAsync([assignment], ct);
-        await scheduleAssignmentRepo.SaveChangesAsync(ct);
-
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<Guid>> CreateGroupAsync(
-        Guid seriesId, CreateGroupRequest request, Guid organizationId, CancellationToken ct = default)
-    {
-        var exists = await lessonSeriesRepo.ExistsAsync(seriesId, organizationId, ct);
-        if (!exists)
-            return Result<Guid>.Fail(new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        var enrollments = await enrollmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var selected = enrollments.Where(e => request.EnrollmentIds.Contains(e.Id)).ToList();
-
-        if (selected.Count != request.EnrollmentIds.Count)
-            return Result<Guid>.Fail(
-                new Error(ErrorCodes.Validation, "Eén of meer inschrijvingen niet gevonden."));
-
-        if (selected.Any(e => e.EnrollmentGroupId.HasValue))
-            return Result<Guid>.Fail(
-                new Error(ErrorCodes.Validation, "Eén of meer inschrijvingen zitten al in een groep."));
-
-        var existingGroupCount = await enrollmentGroupRepo.CountBySeriesAsync(seriesId, organizationId, ct);
-        var groupLetter = (char)('A' + existingGroupCount);
-
-        EnrollmentGroup group = new()
-        {
-            OrganizationId = organizationId,
-            LessonSerieId = seriesId,
-            Name = $"Groep {groupLetter}",
-            LeaderEnrollmentId = selected[0].Id,
-        };
-
-        await enrollmentGroupRepo.AddAsync(group, ct);
-
-        foreach (var enrollment in selected)
-            enrollment.EnrollmentGroupId = group.Id;
-
-        await enrollmentGroupRepo.SaveChangesAsync(ct);
-
-        return Result<Guid>.Ok(group.Id);
-    }
-
-    public async Task<Result<bool>> DissolveGroupAsync(
-        Guid seriesId, Guid groupId, Guid organizationId, CancellationToken ct = default)
-    {
-        var group = await enrollmentGroupRepo.GetByIdAsync(groupId, organizationId, ct);
-        if (group is null || group.LessonSerieId != seriesId)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Groep niet gevonden."));
-
-        // Remove group reference from members
-        foreach (var member in group.Members)
-            member.EnrollmentGroupId = null;
-
-        // Remove any assignments that reference this group
-        var assignments = await scheduleAssignmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var groupAssignments = assignments.Where(a => a.EnrollmentGroupId == groupId).ToList();
-        if (groupAssignments.Count > 0)
-            scheduleAssignmentRepo.RemoveRange(groupAssignments);
-
-        enrollmentGroupRepo.Delete(group);
-        await enrollmentGroupRepo.SaveChangesAsync(ct);
-
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<bool>> ConfirmScheduleAsync(
-        Guid seriesId, Guid organizationId, CancellationToken ct = default)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        if (series is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        if (series.PlanningStatus != PlanningStatus.Planning)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Planning moet eerst gegenereerd worden."));
-
-        var assignments = await scheduleAssignmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var proposedAssignments = assignments.Where(a => a.Status == ScheduleAssignmentStatus.Proposed).ToList();
-
-        if (proposedAssignments.Count == 0)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Er zijn geen voorgestelde toewijzingen om te bevestigen."));
-
-        // Move assignments to AwaitingConfirmation and generate a token per contact point.
-        var expiresAt = DateTime.UtcNow.AddHours(72);
-        var tokens = new List<AssignmentConfirmationToken>();
-        var emailsToSend = new List<(Enrollment recipient, ScheduleAssignment assignment, string rawToken)>();
-        var slotById = series.WeeklyTemplate.ToDictionary(s => s.Id);
-
-        foreach (var assignment in proposedAssignments)
-        {
-            assignment.Status = ScheduleAssignmentStatus.AwaitingConfirmation;
-
-            // Group: token for the leader only. Solo: token for the enrolled individual.
-            Enrollment? recipient = null;
-            if (assignment.EnrollmentGroupId.HasValue && assignment.EnrollmentGroup is not null)
-            {
-                recipient = assignment.EnrollmentGroup.Members
-                    .FirstOrDefault(m => m.Id == assignment.EnrollmentGroup.LeaderEnrollmentId);
-            }
-            else if (assignment.EnrollmentId.HasValue && assignment.Enrollment is not null)
-            {
-                recipient = assignment.Enrollment;
-            }
-
-            if (recipient is null)
-            {
-                logger.LogWarning("Toewijzing {AssignmentId} heeft geen ontvanger — token overgeslagen.", assignment.Id);
-                continue;
-            }
-
-            var rawToken = GenerateRawToken();
-            tokens.Add(new AssignmentConfirmationToken
-            {
-                OrganizationId = organizationId,
-                ScheduleAssignmentId = assignment.Id,
-                EnrollmentId = recipient.Id,
-                TokenHash = HashToken(rawToken),
-                ExpiresAt = expiresAt,
-                Response = ConfirmationResponse.Pending,
-            });
-            emailsToSend.Add((recipient, assignment, rawToken));
-        }
-
-        await tokenRepo.AddRangeAsync(tokens, ct);
-        series.PlanningStatus = PlanningStatus.AwaitingConfirmation;
-        await lessonSeriesRepo.SaveChangesAsync(ct);
-
-        // Send confirmation emails after the state is persisted.
-        var baseUrl = appOptions.Value.ConfirmationBaseUrl.TrimEnd('/');
-        foreach (var (recipient, assignment, rawToken) in emailsToSend)
-        {
-            if (!slotById.TryGetValue(assignment.WeeklyTemplateEntryId, out var slot)) continue;
-
-            try
-            {
-                await emailService.SendScheduleConfirmationAsync(
-                    recipient.StudentEmail,
-                    recipient.StudentName,
-                    series.Name,
-                    slot.DayOfWeek,
-                    slot.StartTime.ToString("HH:mm"),
-                    slot.EndTime.ToString("HH:mm"),
-                    slot.CourtName,
-                    $"{baseUrl}/{rawToken}",
-                    ct);
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex, "E-mail voor toewijzing {AssignmentId} mislukt.", assignment.Id);
-            }
-        }
-
-        logger.LogInformation(
-            "Planning bevestigd voor reeks {SeriesId}: {Count} toewijzingen, {TokenCount} tokens verzonden",
-            seriesId, proposedAssignments.Count, tokens.Count);
-
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<List<NonResponderDto>>> GetNonRespondersAsync(
-        Guid seriesId, Guid organizationId, CancellationToken ct = default)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        if (series is null)
-            return Result<List<NonResponderDto>>.Fail(
-                new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        var tokens = await tokenRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var pending = tokens.Where(t => t.Response == ConfirmationResponse.Pending).ToList();
-        if (pending.Count == 0)
-            return Result<List<NonResponderDto>>.Ok([]);
-
-        var slotById = series.WeeklyTemplate.ToDictionary(s => s.Id);
-        var now = DateTime.UtcNow;
-
-        var list = new List<NonResponderDto>();
-        foreach (var token in pending)
-        {
-            if (!slotById.TryGetValue(token.ScheduleAssignment.WeeklyTemplateEntryId, out var slot))
-                continue;
-
-            var isGroup = token.ScheduleAssignment.EnrollmentGroupId.HasValue
-                && token.ScheduleAssignment.EnrollmentGroup is not null;
-            var groupSize = isGroup ? token.ScheduleAssignment.EnrollmentGroup!.Members.Count : 1;
-
-            list.Add(new NonResponderDto
-            {
-                AssignmentId = token.ScheduleAssignmentId,
-                EnrollmentId = token.EnrollmentId,
-                StudentName = token.Enrollment.StudentName,
-                StudentEmail = token.Enrollment.StudentEmail,
-                StudentPhone = token.Enrollment.StudentPhone,
-                IsGroup = isGroup,
-                GroupSize = groupSize,
-                DayOfWeek = slot.DayOfWeek,
-                StartTime = slot.StartTime.ToString("HH:mm"),
-                EndTime = slot.EndTime.ToString("HH:mm"),
-                CourtName = slot.CourtName,
-                ExpiresAt = token.ExpiresAt,
-                IsExpired = token.ExpiresAt < now,
-            });
-        }
-
-        return Result<List<NonResponderDto>>.Ok(
-            list.OrderBy(r => r.IsExpired ? 0 : 1).ThenBy(r => r.ExpiresAt).ToList());
-    }
-
-    public async Task<Result<bool>> ResendConfirmationEmailAsync(
-        Guid seriesId, Guid assignmentId, Guid organizationId, CancellationToken ct = default)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        if (series is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        var assignment = await scheduleAssignmentRepo.GetByIdAsync(assignmentId, organizationId, ct);
-        if (assignment is null || assignment.LessonSerieId != seriesId)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Toewijzing niet gevonden."));
-
-        if (assignment.Status != ScheduleAssignmentStatus.AwaitingConfirmation)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Enkel toewijzingen in afwachting kunnen opnieuw verstuurd worden."));
-
-        var slot = series.WeeklyTemplate.FirstOrDefault(s => s.Id == assignment.WeeklyTemplateEntryId);
-        if (slot is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Tijdslot niet gevonden."));
-
-        // Determine recipient (group leader for groups, solo enrollee otherwise)
-        Enrollment? recipient = null;
-        if (assignment.EnrollmentGroupId.HasValue && assignment.EnrollmentGroup is not null)
-        {
-            recipient = assignment.EnrollmentGroup.Members
-                .FirstOrDefault(m => m.Id == assignment.EnrollmentGroup.LeaderEnrollmentId);
-        }
-        else if (assignment.EnrollmentId.HasValue && assignment.Enrollment is not null)
-        {
-            recipient = assignment.Enrollment;
-        }
-
-        if (recipient is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.Validation, "Geen ontvanger gevonden voor deze toewijzing."));
-
-        // Invalidate any existing pending token(s) for this assignment by expiring them now
-        var existingTokens = await tokenRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        foreach (var old in existingTokens.Where(t =>
-            t.ScheduleAssignmentId == assignmentId && t.Response == ConfirmationResponse.Pending))
-        {
-            old.ExpiresAt = DateTime.UtcNow;
-        }
-
-        var rawToken = GenerateRawToken();
-        var newToken = new AssignmentConfirmationToken
-        {
-            OrganizationId = organizationId,
-            ScheduleAssignmentId = assignment.Id,
-            EnrollmentId = recipient.Id,
-            TokenHash = HashToken(rawToken),
-            ExpiresAt = DateTime.UtcNow.AddHours(72),
-            Response = ConfirmationResponse.Pending,
-        };
-        await tokenRepo.AddRangeAsync([newToken], ct);
-        await tokenRepo.SaveChangesAsync(ct);
-
-        var baseUrl = appOptions.Value.ConfirmationBaseUrl.TrimEnd('/');
-        try
-        {
-            await emailService.SendScheduleConfirmationAsync(
-                recipient.StudentEmail,
-                recipient.StudentName,
-                series.Name,
-                slot.DayOfWeek,
-                slot.StartTime.ToString("HH:mm"),
-                slot.EndTime.ToString("HH:mm"),
-                slot.CourtName,
-                $"{baseUrl}/{rawToken}",
-                ct);
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, "Resend e-mail voor toewijzing {AssignmentId} mislukt.", assignment.Id);
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "E-mail kon niet verzonden worden. Probeer opnieuw of bevestig handmatig."));
-        }
-
-        logger.LogInformation("Bevestigingsmail opnieuw verstuurd voor toewijzing {AssignmentId}.", assignment.Id);
-        return Result<bool>.Ok(true);
-    }
-
-    public async Task<Result<bool>> AdminConfirmAssignmentAsync(
-        Guid seriesId, Guid assignmentId, Guid organizationId, CancellationToken ct = default)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        if (series is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Lessenreeks niet gevonden."));
-
-        var assignment = await scheduleAssignmentRepo.GetByIdAsync(assignmentId, organizationId, ct);
-        if (assignment is null || assignment.LessonSerieId != seriesId)
-            return Result<bool>.Fail(new Error(ErrorCodes.NotFound, "Toewijzing niet gevonden."));
-
-        if (assignment.Status == ScheduleAssignmentStatus.Confirmed)
-            return Result<bool>.Fail(
-                new Error(ErrorCodes.Validation, "Deze toewijzing is al bevestigd."));
-
-        // Determine payer + group size (for pricing) + enrollees to flip to Confirmed
-        Enrollment? payer;
-        int groupSize;
-        List<Enrollment> affectedEnrollments;
-
-        if (assignment.EnrollmentGroupId.HasValue && assignment.EnrollmentGroup is not null)
-        {
-            payer = assignment.EnrollmentGroup.Members
-                .FirstOrDefault(m => m.Id == assignment.EnrollmentGroup.LeaderEnrollmentId);
-            groupSize = assignment.EnrollmentGroup.Members.Count;
-            affectedEnrollments = assignment.EnrollmentGroup.Members.ToList();
-        }
-        else
-        {
-            payer = assignment.Enrollment;
-            groupSize = 1;
-            affectedEnrollments = payer is not null ? [payer] : [];
-        }
-
-        if (payer is null)
-            return Result<bool>.Fail(new Error(ErrorCodes.Validation, "Geen betaler gevonden voor deze toewijzing."));
-
-        // Record cash payment on the payer's enrollment
-        Payment payment = new()
-        {
-            OrganizationId = organizationId,
-            EnrollmentId = payer.Id,
-            Amount = series.Price * groupSize,
-            Status = PaymentStatus.Paid,
-            Method = PaymentMethod.Cash,
-            PaidAt = DateTime.UtcNow,
-            Description = $"Cash (admin) — {series.Name}",
-        };
-        await paymentRepo.AddAsync(payment, ct);
-
-        // Flip assignment + enrollments + pending tokens
-        assignment.Status = ScheduleAssignmentStatus.Confirmed;
-        foreach (var enrollment in affectedEnrollments)
-            enrollment.Status = EnrollmentStatus.Confirmed;
-
-        var tokens = await tokenRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        foreach (var token in tokens.Where(t =>
-            t.ScheduleAssignmentId == assignmentId && t.Response == ConfirmationResponse.Pending))
-        {
-            token.Response = ConfirmationResponse.Confirmed;
-            token.RespondedAt = DateTime.UtcNow;
-        }
-
-        await paymentRepo.SaveChangesAsync(ct);
-
-        // Check if series is now fully resolved
-        var allTokens = await tokenRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var stillPending = allTokens.Any(t => t.Response == ConfirmationResponse.Pending
-            && t.ExpiresAt >= DateTime.UtcNow);
-        if (!stillPending && series.PlanningStatus == PlanningStatus.AwaitingConfirmation)
-        {
-            series.PlanningStatus = PlanningStatus.Scheduled;
-            await lessonSeriesRepo.SaveChangesAsync(ct);
-        }
-
-        logger.LogInformation("Admin bevestigde toewijzing {AssignmentId} (€{Amount}).", assignment.Id, payment.Amount);
-        return Result<bool>.Ok(true);
-    }
-
-    private static string GenerateRawToken()
-    {
-        Span<byte> bytes = stackalloc byte[32];
-        RandomNumberGenerator.Fill(bytes);
-        return Convert.ToHexString(bytes).ToLowerInvariant();
-    }
-
-    private static string HashToken(string rawToken)
-    {
-        var bytes = SHA256.HashData(Encoding.UTF8.GetBytes(rawToken));
-        return Convert.ToHexString(bytes).ToLowerInvariant();
-    }
-
-    private async Task<Error?> EnsureSlotCapacityAsync(
-        Guid seriesId, Guid organizationId, Guid slotId, int addSize,
-        Guid? excludeAssignmentId, CancellationToken ct)
-    {
-        var series = await lessonSeriesRepo.GetByIdAsync(seriesId, organizationId, ct);
-        var slot = series?.WeeklyTemplate.FirstOrDefault(s => s.Id == slotId);
-        if (slot is null) return null; // caller already validated
-
-        var existing = await scheduleAssignmentRepo.GetBySeriesAsync(seriesId, organizationId, ct);
-        var currentCount = existing
-            .Where(a => a.WeeklyTemplateEntryId == slotId && a.Id != excludeAssignmentId)
-            .Sum(a => a.EnrollmentGroup?.Members.Count ?? 1);
-
-        if (currentCount + addSize > slot.MaxStudents)
-            return new Error(ErrorCodes.Validation,
-                $"Tijdslot heeft geen plaats meer ({currentCount}/{slot.MaxStudents}).");
-
-        return null;
-    }
-
-    private static Dictionary<Guid, SlotPreference> IntersectPreferences(
-        List<Dictionary<Guid, SlotPreference>> memberPrefs)
-    {
-        if (memberPrefs.Count == 0) return new();
-        if (memberPrefs.Count == 1) return memberPrefs[0];
-
-        var allSlotIds = memberPrefs
-            .SelectMany(p => p.Keys)
-            .Distinct();
-
-        var result = new Dictionary<Guid, SlotPreference>();
-
-        foreach (var slotId in allSlotIds)
-        {
-            // Take worst preference among members (most restrictive)
-            var worstPref = SlotPreference.Preferred;
-            var allHave = true;
-
-            foreach (var memberPref in memberPrefs)
-            {
-                if (!memberPref.TryGetValue(slotId, out var pref))
-                {
-                    allHave = false;
-                    break;
-                }
-
-                if (pref == SlotPreference.Unavailable)
-                {
-                    worstPref = SlotPreference.Unavailable;
-                    break;
-                }
-
-                if (pref == SlotPreference.Available && worstPref == SlotPreference.Preferred)
-                    worstPref = SlotPreference.Available;
-            }
-
-            if (!allHave)
-                result[slotId] = SlotPreference.Unavailable;
-            else
-                result[slotId] = worstPref;
-        }
-
-        return result;
     }
 }
